Refector: Refactor the list-books section of the DOM into multiple components. including ListBooks, BookShelf and SingleBook.
Insert the BooksAPI, and called getAll() to retrieve the books on the relevant bookshelves. Passed this through the App component down to SingleBook component through props. Ensure that books only show on their relevant bookshelf. Ensure ability to have multiple books shelves with various names depending on the prop passed, and their shelfFilter also passed as a prop.